# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Planty","version":"0.1.0"},"paths":{"/api/task":{"post":{"tags":["User tasks"],"summary":"Create Task","operationId":"create_task_api_task_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"delete":{"tags":["User tasks"],"summary":"Remove Task","operationId":"remove_task_api_task_delete","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRemoveRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["User tasks"],"summary":"Update Task","operationId":"update_task_api_task_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskUpdateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/task/by_date":{"get":{"tags":["User tasks"],"summary":"Get Tasks By Date","operationId":"get_tasks_by_date_api_task_by_date_get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"not_before","in":"query","required":true,"schema":{"type":"string","format":"date","title":"Not Before"}},{"name":"not_after","in":"query","required":true,"schema":{"type":"string","format":"date","title":"Not After"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponse"}},"title":"Response Get Tasks By Date Api Task By Date Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/task/search":{"get":{"tags":["User tasks"],"summary":"Get Tasks By Search Query","operationId":"get_tasks_by_search_query_api_task_search_get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponse"},"title":"Response Get Tasks By Search Query Api Task Search Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/task/move":{"post":{"tags":["User tasks"],"summary":"Move Task","operationId":"move_task_api_task_move_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskMoveRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/task/toggle_completed":{"post":{"tags":["User tasks"],"summary":"Toggle Task Completed","operationId":"toggle_task_completed_api_task_toggle_completed_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToggleCompletedRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/task/toggle_archived":{"post":{"tags":["User tasks"],"summary":"Toggle Task Archived","operationId":"toggle_task_archived_api_task_toggle_archived_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskToggleArchivedRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/task/attachment":{"post":{"tags":["User tasks"],"summary":"Get Attachment Uploading Info","description":"This endpoint allows the frontend to obtain a pre-signed POST URL along with the required fields for uploading an attachment to an S3-compatible storage. The frontend is responsible for encrypting the file client-side using AES-128 CBC with the provided key and IV. After encryption, the frontend directly uploads the file to the S3 storage using the pre-signed URL and fields. \n\nThe frontend can include the 'Content-Disposition' header in the upload request to specify the file name, ensuring that the file is downloaded later with the correct name.\n\nThe approach with client-side encryption allows using even non-trusted S3 Storage providers for user files.","operationId":"get_attachment_uploading_info_api_task_attachment_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestAttachmentUpload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttachmentUploadInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/task/{task_id}/attachment/{attachment_id}":{"delete":{"tags":["User tasks"],"summary":"Remove Attachment","operationId":"remove_attachment_api_task__task_id__attachment__attachment_id__delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Task Id"}},{"name":"attachment_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Attachment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/archived":{"get":{"tags":["User tasks"],"summary":"Get Archived Tasks","operationId":"get_archived_tasks_api_tasks_archived_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArchivedTasksResponse"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/section":{"post":{"tags":["User tasks"],"summary":"Create Section","operationId":"create_section_api_section_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["User tasks"],"summary":"Patch Section","operationId":"patch_section_api_section_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionUpdateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/section/{section_id}":{"get":{"tags":["User tasks"],"summary":"Get Section","operationId":"get_section_api_section__section_id__get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"section_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Section Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/section/move":{"post":{"tags":["User tasks"],"summary":"Move Section","operationId":"move_section_api_section_move_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionMoveRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/sections":{"get":{"tags":["User tasks"],"summary":"Get Sections","operationId":"get_sections_api_sections_get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"leaves_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Leaves Only"}},{"name":"as_tree","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"As Tree"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SectionResponse"},"title":"Response Get Sections Api Sections Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/section/shuffle":{"post":{"tags":["User tasks"],"summary":"Shuffle Section","operationId":"shuffle_section_api_section_shuffle_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShuffleSectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/auth/login":{"post":{"tags":["auth"],"summary":"Auth:Db Cookie.Login","operationId":"auth_db_cookie_login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_auth_db_cookie_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"LOGIN_BAD_CREDENTIALS":{"summary":"Bad credentials or the user is inactive.","value":{"detail":"LOGIN_BAD_CREDENTIALS"}},"LOGIN_USER_NOT_VERIFIED":{"summary":"The user is not verified.","value":{"detail":"LOGIN_USER_NOT_VERIFIED"}}}}}},"204":{"description":"No Content"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth"],"summary":"Auth:Db Cookie.Logout","operationId":"auth_db_cookie_logout_api_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"Missing token or inactive user."},"204":{"description":"No Content"}},"security":[{"APIKeyCookie":[]}]}},"/api/auth/register":{"post":{"tags":["auth"],"summary":"Register:Register","operationId":"register_register_api_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"REGISTER_USER_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"REGISTER_USER_ALREADY_EXISTS"}},"REGISTER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"REGISTER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/me":{"get":{"tags":["auth"],"summary":"Users:Current User","operationId":"users_current_user_api_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."}},"security":[{"APIKeyCookie":[]}]},"patch":{"tags":["auth"],"summary":"Users:Patch Current User","operationId":"users_patch_current_user_api_auth_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorModel"},"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyCookie":[]}]}},"/api/auth/{id}":{"get":{"tags":["auth"],"summary":"Users:User","operationId":"users_user_api_auth__id__get","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["auth"],"summary":"Users:Patch User","operationId":"users_patch_user_api_auth__id__patch","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"400":{"content":{"application/json":{"examples":{"UPDATE_USER_EMAIL_ALREADY_EXISTS":{"summary":"A user with this email already exists.","value":{"detail":"UPDATE_USER_EMAIL_ALREADY_EXISTS"}},"UPDATE_USER_INVALID_PASSWORD":{"summary":"Password validation failed.","value":{"detail":{"code":"UPDATE_USER_INVALID_PASSWORD","reason":"Password should beat least 3 characters"}}}},"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Bad Request"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["auth"],"summary":"Users:Delete User","operationId":"users_delete_user_api_auth__id__delete","security":[{"APIKeyCookie":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"204":{"description":"Successful Response"},"401":{"description":"Missing token or inactive user."},"403":{"description":"Not a superuser."},"404":{"description":"The user does not exist."},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ArchivedTasksResponse":{"properties":{"title":{"type":"string","title":"Title","default":"Archived tasks"},"tasks":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Tasks"}},"additionalProperties":false,"type":"object","required":["tasks"],"title":"ArchivedTasksResponse"},"AttachmentResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"aes_key_b64":{"type":"string","title":"Aes Key B64"},"aes_iv_b64":{"type":"string","title":"Aes Iv B64"},"s3_file_key":{"type":"string","title":"S3 File Key"},"task_id":{"type":"string","format":"uuid","title":"Task Id"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"url":{"type":"string","title":"Url"}},"type":"object","required":["aes_key_b64","aes_iv_b64","s3_file_key","task_id","added_at","url"],"title":"AttachmentResponse"},"AttachmentUploadInfo":{"properties":{"post_url":{"type":"string","title":"Post Url"},"post_fields":{"type":"object","title":"Post Fields"}},"additionalProperties":false,"type":"object","required":["post_url","post_fields"],"title":"AttachmentUploadInfo"},"Body_auth_db_cookie_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_auth_db_cookie_login_api_auth_login_post"},"ErrorModel":{"properties":{"detail":{"anyOf":[{"type":"string"},{"additionalProperties":{"type":"string"},"type":"object"}],"title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorModel"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RecurrenceInfo":{"properties":{"period":{"type":"integer","title":"Period"},"type":{"type":"string","enum":["days","weeks","months","years"],"title":"Type"},"flexible_mode":{"type":"boolean","title":"Flexible Mode"}},"type":"object","required":["period","type","flexible_mode"],"title":"RecurrenceInfo"},"RequestAttachmentUpload":{"properties":{"task_id":{"type":"string","format":"uuid","title":"Task Id"},"aes_key_b64":{"type":"string","title":"Aes Key B64"},"aes_iv_b64":{"type":"string","title":"Aes Iv B64"}},"additionalProperties":false,"type":"object","required":["task_id","aes_key_b64","aes_iv_b64"],"title":"RequestAttachmentUpload"},"SectionCreateRequest":{"properties":{"title":{"type":"string","title":"Title"},"parent_id":{"type":"string","format":"uuid","title":"Parent Id"}},"additionalProperties":false,"type":"object","required":["title","parent_id"],"title":"SectionCreateRequest"},"SectionCreateResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"}},"additionalProperties":false,"type":"object","required":["id"],"title":"SectionCreateResponse"},"SectionMoveRequest":{"properties":{"section_id":{"type":"string","format":"uuid","title":"Section Id"},"to_parent_id":{"type":"string","format":"uuid","title":"To Parent Id"},"index":{"type":"integer","minimum":0.0,"title":"Index"}},"additionalProperties":false,"type":"object","required":["section_id","to_parent_id","index"],"title":"SectionMoveRequest"},"SectionResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"parent_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Parent Id"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"subsections":{"items":{"$ref":"#/components/schemas/SectionResponse"},"type":"array","title":"Subsections"},"tasks":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Tasks"}},"additionalProperties":false,"type":"object","required":["id","title","parent_id","added_at","subsections","tasks"],"title":"SectionResponse"},"SectionUpdateRequest":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"}},"additionalProperties":false,"type":"object","required":["id"],"title":"SectionUpdateRequest"},"SectionUpdateResponse":{"properties":{"section":{"$ref":"#/components/schemas/SectionResponse"}},"additionalProperties":false,"type":"object","required":["section"],"title":"SectionUpdateResponse"},"ShuffleSectionRequest":{"properties":{"section_id":{"type":"string","format":"uuid","title":"Section Id"}},"additionalProperties":false,"type":"object","required":["section_id"],"title":"ShuffleSectionRequest"},"TaskCreateRequest":{"properties":{"section_id":{"type":"string","format":"uuid","title":"Section Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"due_to":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Due To"},"recurrence":{"anyOf":[{"$ref":"#/components/schemas/RecurrenceInfo"},{"type":"null"}]}},"additionalProperties":false,"type":"object","required":["section_id","title"],"title":"TaskCreateRequest"},"TaskCreateResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"}},"additionalProperties":false,"type":"object","required":["id"],"title":"TaskCreateResponse"},"TaskMoveRequest":{"properties":{"task_id":{"type":"string","format":"uuid","title":"Task Id"},"section_to_id":{"type":"string","format":"uuid","title":"Section To Id"},"index":{"type":"integer","minimum":0.0,"title":"Index"}},"additionalProperties":false,"type":"object","required":["task_id","section_to_id","index"],"title":"TaskMoveRequest"},"TaskRemoveRequest":{"properties":{"task_id":{"type":"string","format":"uuid","title":"Task Id"}},"additionalProperties":false,"type":"object","required":["task_id"],"title":"TaskRemoveRequest"},"TaskResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"section_id":{"type":"string","format":"uuid","title":"Section Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"is_completed":{"type":"boolean","title":"Is Completed"},"is_archived":{"type":"boolean","title":"Is Archived"},"added_at":{"type":"string","format":"date-time","title":"Added At"},"due_to":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Due To"},"recurrence":{"anyOf":[{"$ref":"#/components/schemas/RecurrenceInfo"},{"type":"null"}]},"attachments":{"items":{"$ref":"#/components/schemas/AttachmentResponse"},"type":"array","title":"Attachments"}},"additionalProperties":false,"type":"object","required":["id","section_id","title","description","content","is_completed","is_archived","added_at","due_to","recurrence","attachments"],"title":"TaskResponse"},"TaskToggleArchivedRequest":{"properties":{"task_id":{"type":"string","format":"uuid","title":"Task Id"}},"additionalProperties":false,"type":"object","required":["task_id"],"title":"TaskToggleArchivedRequest"},"TaskToggleCompletedRequest":{"properties":{"task_id":{"type":"string","format":"uuid","title":"Task Id"},"auto_archive":{"type":"boolean","title":"Auto Archive","default":true}},"additionalProperties":false,"type":"object","required":["task_id"],"title":"TaskToggleCompletedRequest"},"TaskUpdateRequest":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"due_to":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Due To"}},"additionalProperties":false,"type":"object","required":["id"],"title":"TaskUpdateRequest"},"TaskUpdateResponse":{"properties":{"task":{"$ref":"#/components/schemas/TaskResponse"}},"additionalProperties":false,"type":"object","required":["task"],"title":"TaskUpdateResponse"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active","default":true},"is_superuser":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Superuser","default":false},"is_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Verified","default":false}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active","default":true},"is_superuser":{"type":"boolean","title":"Is Superuser","default":false},"is_verified":{"type":"boolean","title":"Is Verified","default":false}},"type":"object","required":["id","email"],"title":"UserRead"},"UserUpdate":{"properties":{"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"is_superuser":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Superuser"},"is_verified":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Verified"}},"type":"object","title":"UserUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyCookie":{"type":"apiKey","in":"cookie","name":"fastapiusersauth"}}}}